test: 중복해서 사용하는 스니펫들을 공통으로 관리하라
