--- src/test/java/prgrms/project/stuti/domain/studygroup/controller/CommonStudyDomainDescriptor.java
@@ -1,2 +1,0 @@
-package prgrms.project.stuti.domain.studygroup.controller;public class CommonStudyDomainDescriptor {
-}
